+++ src/main/java/deboni/potatologistics/blocks/entities/TileEntityCoil.java
@@ -1,6 +1,8 @@
 package deboni.potatologistics.blocks.entities;
 
 import net.minecraft.core.block.entity.TileEntity;
+import net.minecraft.core.net.packet.Packet;
+import net.minecraft.core.net.packet.Packet140TileEntityData;
 import sunsetsatellite.energyapi.impl.TileEntityEnergyConductor;
 import sunsetsatellite.sunsetutils.util.Connection;
 import sunsetsatellite.sunsetutils.util.Direction;
@@ -26,6 +28,10 @@ public class TileEntityCoil extends TileEntityEnergyConductor {
             int energyProduced = engine.consumeEnergy();
             modifyEnergy(energyProduced);
         }
+    }
+    @Override
+    public Packet getDescriptionPacket() {
+        return new Packet140TileEntityData(this);
     }
