+++ src/main/java/deboni/potatologistics/blocks/entities/TileEntityPipe.java
@@ -8,6 +8,8 @@ import net.minecraft.core.block.BlockChest;
 import net.minecraft.core.block.entity.TileEntity;
 import net.minecraft.core.block.entity.TileEntityChest;
 import net.minecraft.core.item.ItemStack;
+import net.minecraft.core.net.packet.Packet;
+import net.minecraft.core.net.packet.Packet140TileEntityData;
 import net.minecraft.core.player.inventory.IInventory;
 import net.minecraft.core.util.helper.Direction;
 import sunsetsatellite.sunsetutils.util.Connection;
@@ -251,5 +253,9 @@ public class TileEntityPipe extends TileEntity {
                 stack.timer--;
             }
         }
+    }
+    @Override
+    public Packet getDescriptionPacket() {
+        return new Packet140TileEntityData(this);
     }
 }
